requirements.txt
# Core pipeline & query
numpy<2
pandas>=2.0
opencv-python-headless>=4.8
scenedetect[opencv]>=0.6
google-genai>=1.0
openpyxl>=3.1
boto3>=1.28

# Face / vision
insightface>=0.7
tqdm>=4.65
scikit-learn>=1.3
onnxruntime>=1.16

# Embeddings & vector DB
sentence-transformers[torch]>=2.2
pymilvus>=2.3
psycopg2-binary>=2.9

# Optional: .env for local overrides (see utils/aud_db_utils.py)
python-dotenv>=1.0
